собрал в кучу компоненты и хуки мап
